Fix preview, and extra draw during stop
Journal preview image was blank.

Also, when the stop button was pressed, or ctrl+q, the canvas did redraw
once.

During port to GTK+ 3, the canvas Gtk.DrawingArea draw signal handler
was not changed to draw on the provided cairo.Context

The handler is used by activity.Activity to draw on the preview surface.